Move PipelineChannel memory impl
