+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/java/org/apache/shardingsphere/data/pipeline/core/ingest/channel/memory/ManualBitSet.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.core.ingest.channel.distribution;
+package org.apache.shardingsphere.data.pipeline.core.ingest.channel.memory;
 
 import java.util.BitSet;
 import java.util.LinkedList;
